Comment&reply rework

+++ LearningEnhancementSystem/src/java/Servlets/OneModule.java
@@ -66,8 +66,9 @@ public class OneModule extends SuperServlet {
                     if (paramap.get("delete")[0].equals("TRUE")) {
                     String comid = request.getParameter("comment_id");
                     int commId = Integer.parseInt(comid);
-                    cdb.deleteComment(commId);
                     crdb.deleteAll(comid);
+                    cdb.deleteComment(commId);
+                    
                     }
                 } if (paramap.containsKey("deleteR")) {
                     String repid = request.getParameter("reply_id");
@@ -103,47 +104,40 @@ public class OneModule extends SuperServlet {
             List<Comment> commentList = cdb.getComments();
             List<CommentReply> replyList = crdb.getCommentReplys();
 
-            out.println("<p>");
-            out.println("<div class=\"jumbotron\">");
-            out.println("<div class=\"container\">");
-            out.println("<button class=\"btn btn-link\" data-toggle=\"collapse\" data-target=\"#collapse\" aria-expanded=\"true\" aria-controls=\"collapse\">");
-            out.println("<h4 class=\"display-4\">Kommentarer</h4>");
-            out.println("<hr class=\"my-4\">");
-            out.println("</button>");
-            out.println("<div class=\"collapse\" id=\"collapse\">");
-            out.println("<div class=\"card-body\">");
+            bst.collapseTop(out);
             for (Comment comment : commentList){
                 String commentId = comment.getCommentId();
                 String commentText = comment.getCommentText();
                 String commentUserName = comment.getUserName();
+                String commentUserId = comment.getUserId();
                 String commentModuleId = comment.getModuleId();
                 if (commentModuleId.equals(id)){
                     out.println("<p>" + commentText + "</p>");
                     out.println("<p>" + commentUserName + "</p>");
+                    if (user.getUserIsTeacher()||(user.getUserId().equals(commentUserId))){
                     deleteComment(out,commentModuleId,commentId);
+                    }
                     for (CommentReply reply : replyList){
                         String replyId = reply.getReplyId();
                         String replyText = reply.getReplyText();
                         String replyUserName = reply.getUserName();
+                        String replyUserId = reply.getUserId();
                         String replyCommentId = reply.getCommentId();
                         if (commentId.equals(replyCommentId)){
                             out.println("<hr class=\"my-4\">");
                             out.println("<p style=\"margin-left:2.5em;\">" + replyText + "</p>");
                             out.println("<p style=\"margin-left:2.5em;\">" + replyUserName + "</p>");
+                            if (user.getUserIsTeacher()||(user.getUserId().equals(replyUserId))){
                             deleteReply(out,commentModuleId,replyId);
+                            }
                         }
                     }
                     addReply(out,commentModuleId,commentId);
                     out.println("<hr class=\"my-4\">");
                 }
             }
-            //cdb.printComments(mId,out);
-
-           // cdb.addCommentForm(out,mId);
-          
             addComment(out,request);
-            out.println("</div>");
-            out.println("</div>");
+            bst.collapseBottom(out);
             bst.containerClose(out);
             bst.bootstrapFooter(out); 
         }
